Add identity service with Identity Server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -53,12 +54,13 @@ namespace AuctionService
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: add current user as seller
-            auction.Seller = "test";
+
+            auction.Seller = User.Identity!.Name;
 
             _context.Auctions.Add(auction);
 
@@ -76,6 +78,7 @@ namespace AuctionService
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -84,7 +87,7 @@ namespace AuctionService
 
             if (auction == null) return NotFound();
 
-            //TODO: check seller == username
+            if (auction.Seller != User.Identity!.Name) return Forbid();
 
             auction!.Item!.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -100,6 +103,7 @@ namespace AuctionService
             return BadRequest();
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -107,7 +111,7 @@ namespace AuctionService
 
             if (auction == null) return NotFound();
 
-            //TODO: check seller == username
+            if (auction.Seller != User.Identity!.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
